chore: bump stub_gen crates
