+++ pyproject.toml
@@ -17,5 +17,5 @@ authors = [
 description = "OxAPY is http server for python build in rust"
 dependencies = ["orjson>=3.11.3"]
 [tool.maturin]
-python-source="oxapy"
+python-source="."
 features = ["pyo3/extension-module"]
